% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protected-cols.R
\name{gopher_protected_cols_default}
\alias{gopher_protected_cols_default}
\title{Default protected (non-editable) columns per table}
\usage{
gopher_protected_cols_default()
}
\description{
Default protected (non-editable) columns per table
}
\keyword{internal}
