% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gopher_table_names}
\alias{gopher_table_names}
\title{List user-facing table (and optionally view) names in the gopher DB}
\usage{
gopher_table_names(
  con,
  include_views = FALSE,
  include_internal = FALSE,
  prefix = NULL
)
}
\arguments{
\item{con}{A DBI connection.}

\item{include_views}{Logical; include views (TRUE) or tables only (FALSE).}

\item{include_internal}{Logical; include sqlite_* internal tables/views.}

\item{prefix}{Optional character scalar. If provided, only return names that
start with this prefix (e.g. "gopher_" or "gopher."). Set NULL to disable.}
}
\value{
Character vector of table/view names.
}
\description{
List user-facing table (and optionally view) names in the gopher DB
}
