% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{read_bundle}
\alias{read_bundle}
\title{Read a bundled .xlsx file}
\usage{
read_bundle(
  xlsx,
  tables = NULL,
  db_path = getOption("gopher.db_path"),
  db_file = getOption("gopher.db_file")
)
}
\arguments{
\item{xlsx}{path to the bundle.xlsx file}

\item{tables}{list of table to include (default: NULL for all tables)}

\item{db_path}{db path}

\item{db_file}{db file}
}
\value{
a list
}
\description{
Read a bundled .xlsx file
}
